\name{camog-package}
\alias{camog-package}
\alias{camog}
\docType{package}
\title{
\packageTitle{camog}
}
\description{
\packageDescription{camog}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{camog}
\packageIndices{camog}

\preformatted{csvload(filename)}
}
\author{
\packageAuthor{camog}

Maintainer: \packageMaintainer{camog}
}
\keyword{ package }
\keyword{ file }
